server: Add an API to get the file descriptor for a client This adds an API to get the file descriptor for a client. The client file descriptor can be used for a wayland compositor to validate a request from a client if there are any additional information provided from the client's file descriptor.
For instance, this will be helpful in some linux distributions, in which SELinux
or SMACK is enabled. In those environments, each file (including socket) will have
each security contexts in its inode as xattr member variable. A wayland compositor
can validate a client request by getting the file descriptor of the client and
by checking the security contexts associated with the file descriptor.

Change-Id: I563f3941454d87f15a4ffe58786f1fc6cd1d3b24
Signed-off-by: Sung-Jin Park <sj76.park@samsung.com>